+++ sentry/src/main/java/io/sentry/NoOpSpan.java
@@ -70,7 +70,7 @@ public final class NoOpSpan implements ISpan {
 
   @Override
   public @NotNull SpanContext getSpanContext() {
-    return new SpanContext(SentryId.EMPTY_ID, SpanId.EMPTY_ID, null, null);
+    return new SpanContext(SentryId.EMPTY_ID, SpanId.EMPTY_ID, "op", null, null);
   }
 
   @Override