+++ sentry/src/test/java/io/sentry/ScopeTest.kt
@@ -86,7 +86,7 @@ class ScopeTest {
         scope.setTag("tag", "tag")
         scope.setExtra("extra", "extra")
 
-        val transaction = SentryTransaction("transaction-name")
+        val transaction = SentryTransaction("transaction-name", "op")
         scope.setTransaction(transaction)
 
         val attachment = Attachment("path/log.txt")
@@ -147,7 +147,7 @@ class ScopeTest {
         scope.level = SentryLevel.FATAL
         user.id = "456"
 
-        scope.setTransaction(SentryTransaction("newTransaction"))
+        scope.setTransaction(SentryTransaction("newTransaction", "op"))
 
         // because you can only set a new list to scope
         val newFingerprints = mutableListOf("def", "ghf")
@@ -189,7 +189,7 @@ class ScopeTest {
     fun `clear scope resets scope to default state`() {
         val scope = Scope(SentryOptions())
         scope.level = SentryLevel.WARNING
-        scope.setTransaction(SentryTransaction(""))
+        scope.setTransaction(SentryTransaction("", "op"))
         scope.user = User()
         scope.fingerprint = mutableListOf("finger")
         scope.addBreadcrumb(Breadcrumb())
@@ -607,7 +607,7 @@ class ScopeTest {
     @Test
     fun `Scope getTransaction returns the transaction if there is no active span`() {
         val scope = Scope(SentryOptions())
-        val transaction = SentryTransaction("name")
+        val transaction = SentryTransaction("name", "op")
         scope.setTransaction(transaction)
         assertEquals(transaction, scope.span)
     }
@@ -615,7 +615,7 @@ class ScopeTest {
     @Test
     fun `Scope getTransaction returns the current span if there is an unfinished span`() {
         val scope = Scope(SentryOptions())
-        val transaction = SentryTransaction("name")
+        val transaction = SentryTransaction("name", "op")
         scope.setTransaction(transaction)
         val span = transaction.startChild("op")
         assertEquals(span, scope.span)
@@ -624,7 +624,7 @@ class ScopeTest {
     @Test
     fun `Scope getTransaction returns the current span if there is a finished span`() {
         val scope = Scope(SentryOptions())
-        val transaction = SentryTransaction("name")
+        val transaction = SentryTransaction("name", "op")
         scope.setTransaction(transaction)
         val span = transaction.startChild("op")
         span.finish()
@@ -634,7 +634,7 @@ class ScopeTest {
     @Test
     fun `Scope getTransaction returns the latest span if there is a list of active span`() {
         val scope = Scope(SentryOptions())
-        val transaction = SentryTransaction("name")
+        val transaction = SentryTransaction("name", "op")
         scope.setTransaction(transaction)
         val span = transaction.startChild("op")
         val innerSpan = span.startChild("op")
@@ -684,7 +684,7 @@ class ScopeTest {
     @Test
     fun `when transaction is started, sets transaction name on the transaction object`() {
         val scope = Scope(SentryOptions())
-        val sentryTransaction = SentryTransaction("transaction-name")
+        val sentryTransaction = SentryTransaction("transaction-name", "op")
         scope.setTransaction(sentryTransaction)
         assertEquals("transaction-name", scope.transactionName)
         scope.setTransaction("new-name")
@@ -697,7 +697,7 @@ class ScopeTest {
     fun `when transaction is set after transaction name is set, clearing transaction does not bring back old transaction name`() {
         val scope = Scope(SentryOptions())
         scope.setTransaction("transaction-a")
-        val sentryTransaction = SentryTransaction("transaction-name")
+        val sentryTransaction = SentryTransaction("transaction-name", "op")
         scope.setTransaction(sentryTransaction)
         assertEquals("transaction-name", scope.transactionName)
         scope.clearTransaction()