+++ sentry/src/test/java/io/sentry/SentryClientTest.kt
@@ -746,7 +746,7 @@ class SentryClientTest {
     @Test
     fun `transactions are sent using connection`() {
         val sut = fixture.getSut()
-        sut.captureTransaction(SentryTransaction("a-transaction"), mock(), null)
+        sut.captureTransaction(SentryTransaction("a-transaction", "op"), mock(), null)
         verify(fixture.transport).send(check {
             val transaction = it.items.first().getTransaction(fixture.sentryOptions.serializer)
             assertNotNull(transaction)
@@ -756,7 +756,7 @@ class SentryClientTest {
 
     @Test
     fun `when captureTransaction with attachments`() {
-        val transaction = SentryTransaction("a-transaction")
+        val transaction = SentryTransaction("a-transaction", "op")
         fixture.getSut().captureTransaction(transaction, createScopeWithAttachments(), null)
 
         verifyAttachmentsInEnvelope(transaction.eventId)
@@ -764,7 +764,7 @@ class SentryClientTest {
 
     @Test
     fun `when captureTransaction with attachments not added to transaction`() {
-        val transaction = SentryTransaction("a-transaction")
+        val transaction = SentryTransaction("a-transaction", "op")
         val scope = createScopeWithAttachments()
         scope.addAttachment(Attachment("hello".toByteArray(), "application/octet-stream"))
         fixture.getSut().captureTransaction(transaction, scope, null)
@@ -777,7 +777,7 @@ class SentryClientTest {
         val event = SentryEvent()
         val sut = fixture.getSut()
         val scope = createScope()
-        val transaction = SentryTransaction("name")
+        val transaction = SentryTransaction("name", "op")
         scope.setTransaction(transaction)
         transaction.finish()
         sut.captureEvent(event, scope)
@@ -790,7 +790,7 @@ class SentryClientTest {
         val event = SentryEvent()
         val sut = fixture.getSut()
         val scope = createScope()
-        val transaction = SentryTransaction("name")
+        val transaction = SentryTransaction("name", "op")
         scope.setTransaction(transaction)
         val span = transaction.startChild("op")
         sut.captureEvent(event, scope)
@@ -803,7 +803,7 @@ class SentryClientTest {
         fixture.sentryOptions.release = "optionsRelease"
         fixture.sentryOptions.environment = "optionsEnvironment"
         val sut = fixture.getSut()
-        val transaction = SentryTransaction("name")
+        val transaction = SentryTransaction("name", "op")
         sut.captureTransaction(transaction)
         assertEquals("optionsRelease", transaction.release)
         assertEquals("optionsEnvironment", transaction.environment)
@@ -814,7 +814,7 @@ class SentryClientTest {
         fixture.sentryOptions.release = "optionsRelease"
         fixture.sentryOptions.environment = "optionsEnvironment"
         val sut = fixture.getSut()
-        val transaction = SentryTransaction("name")
+        val transaction = SentryTransaction("name", "op")
         transaction.release = "transactionRelease"
         transaction.environment = "transactionEnvironment"
         sut.captureTransaction(transaction)
@@ -826,7 +826,7 @@ class SentryClientTest {
     fun `when transaction does not have tags, and tags are set on options, options values are applied to transactions`() {
         fixture.sentryOptions.setTag("tag1", "value1")
         val sut = fixture.getSut()
-        val transaction = SentryTransaction("name")
+        val transaction = SentryTransaction("name", "op")
         sut.captureTransaction(transaction)
         assertEquals(mapOf("tag1" to "value1"), transaction.tags)
     }
@@ -836,7 +836,7 @@ class SentryClientTest {
         fixture.sentryOptions.setTag("tag1", "value1")
         fixture.sentryOptions.setTag("tag2", "value2")
         val sut = fixture.getSut()
-        val transaction = SentryTransaction("name")
+        val transaction = SentryTransaction("name", "op")
         transaction.setTag("tag3", "value3")
         transaction.setTag("tag2", "transaction-tag")
         sut.captureTransaction(transaction)